Preserve local TData host stream contract
